tests/wpt/tests/scroll-animations/css/scroll-timeline-name-computed.html

<!DOCTYPE html>
<link rel="help" href="https://drafts.csswg.org/scroll-animations-1/#scroll-timeline-name">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<style>
  #outer { scroll-timeline-name: --foo; }
  #target { scroll-timeline-name: --bar; }
</style>
<div id="outer">
  <div id="target"></div>
</div>
<script>
test_computed_value('scroll-timeline-name', 'initial', 'none');
test_computed_value('scroll-timeline-name', 'inherit', '--foo');
test_computed_value('scroll-timeline-name', 'unset', 'none');
test_computed_value('scroll-timeline-name', 'revert', 'none');
test_computed_value('scroll-timeline-name', 'none');
test_computed_value('scroll-timeline-name', '--foo');
test_computed_value('scroll-timeline-name', '--foo, --bar');
test_computed_value('scroll-timeline-name', '--bar, --foo');
test_computed_value('scroll-timeline-name', '--a, --b, --c, --D, --e');
test_computed_value('scroll-timeline-name', 'none, none');
test_computed_value('scroll-timeline-name', '--a, --b, --c, none, --d, --e');

test(() => {
  let style = getComputedStyle(document.getElementById('target'));
  assert_not_equals(Array.from(style).indexOf('scroll-timeline-name'), -1);
}, 'The scroll-timeline-name property shows up in CSSStyleDeclaration enumeration');

test(() => {
  let style = document.getElementById('target').style;
  assert_not_equals(style.cssText.indexOf('scroll-timeline-name'), -1);
}, 'The scroll-timeline-name property shows up in CSSStyleDeclaration.cssText');

</script>
